Move get_index_file_location_and_download_if_missing to libcommon
